chore: 플로깅 포인트 조회 수정

+++ src/main/java/com/innocamp/sanmolong/api/dto/SliderInfoResponseDto.java
@@ -12,5 +12,5 @@ public class SliderInfoResponseDto {
     List<WeatherDto> weathers;
     List<FireDto> fires;
     List<Badge> badges;
-    int plogging;
+    long plogging;
 }
